+++ saviSimulation/savi/jason_processing/ROBOT_model.java
@@ -55,45 +55,45 @@ public class ROBOT_model extends PApplet {
 	int simTimeDelta; // discrete-time step (in seconds)
 	boolean simPaused;// simulation paused or not
 
-  List<WorldObject> objects = new ArrayList<WorldObject>();//List of world objects  
-  List<Threat> threats = new ArrayList<Threat>(); //List of threats
-  List <UAS> UAS_list = new ArrayList<UAS>(); //List of UAS 
+	List<WorldObject> objects = new ArrayList<WorldObject>();//List of world objects  
+	List<Threat> threats = new ArrayList<Threat>(); //List of threats
+	List <UAS> UAS_list = new ArrayList<UAS>(); //List of UAS 
 
-  JasonMAS jasonAgents; // the BDI agent
+	JasonMAS jasonAgents; // the BDI agent
 
-  Button playButton,stopButton,pauseButton;
-  PShape robot,tree,house,threat,play,pause,restart;
+	Button playButton,stopButton,pauseButton;
+	PShape robot,tree,house,threat,play,pause,restart;
 
-  public void settings() { size(X_PIXELS,Y_PIXELS, P2D);  smooth(8); } // let's assume a 2D environment
+	public void settings() { size(X_PIXELS,Y_PIXELS, P2D);  smooth(8); } // let's assume a 2D environment
 
-  public static void main(String[] passedArgs) {
-	  String[] appletArgs = new String[] { "savi.jason_processing.ROBOT_model" };
-	  if (passedArgs != null) {
-		  PApplet.main(concat(appletArgs, passedArgs));
-	  } else {
-		  PApplet.main(appletArgs);
-	  }
-  }
+	public static void main(String[] passedArgs) {
+		String[] appletArgs = new String[] { "savi.jason_processing.ROBOT_model" };
+		if (passedArgs != null) {
+			PApplet.main(concat(appletArgs, passedArgs));
+		} else {
+			PApplet.main(appletArgs);
+		}
+	}
 
 /************* Main setup() ***********************/
 // Main system initialization function
 // called by Processing once on startup
 //************************************************/
 public void setup() {
- /***LOAD SIM PARAMETERS ****/
-  try {
-	  String filePath = new File("").getAbsolutePath();
-	  filePath = filePath + "/config.cfg";
-	  System.out.println(filePath);
-	  File inFile = new File(filePath);
+	/***LOAD SIM PARAMETERS ****/
+	try {
+		String filePath = new File("").getAbsolutePath();
+		filePath = filePath + "/config.cfg";
+		System.out.println(filePath);
+		File inFile = new File(filePath);
 		in = new FileInputStream(inFile);
 		modelProps.load(in);
 		}
-		catch (FileNotFoundException  e) {
-			System.out.println("File not found");
-		}catch (Exception  e) {
-			System.out.println("Exception occurred");
-		}
+	catch (FileNotFoundException  e) {
+		System.out.println("File not found");
+	}catch (Exception  e) {
+		System.out.println("Exception occurred");
+	}
 	NUMBER_TREES = Integer.parseInt(modelProps.getProperty("NUMBER_TREES"));
 	NUMBER_HOUSES = Integer.parseInt(modelProps.getProperty("NUMBER_HOUSES"));
 	NUMBER_THREATS = Integer.parseInt(modelProps.getProperty("NUMBER_TREES"));
@@ -107,7 +107,7 @@ public void setup() {
 	WIFI_PERCEPTION_DISTANCE = Integer.parseInt(modelProps.getProperty("WIFI_PERCEPTION_DISTANCE"));
 	NUMBER_UAS = Integer.parseInt(modelProps.getProperty("NUMBER_UAS"));
 	
-  // let's assume a 2D environment
+	// let's assume a 2D environment
 	// Initialization code goes here
 	simTime = 0;      // seconds
 	simTimeDelta = 1; // seconds
@@ -149,29 +149,29 @@ public void setup() {
 	}          
   
   // smoother rendering (optional)
-		frameRate(FRAME_RATE); // max 60 draw() calls per real second. (Can make it a larger value for the simulation to go faster)
-		// At simTimeDelta=1 this means 1 second
-		// of real time = 1 minute of sim time.
-		//If the processor is not fast enough to maintain the specified rate, the frame rate will not be achieved 
+	frameRate(FRAME_RATE); // max 60 draw() calls per real second. (Can make it a larger value for the simulation to go faster)
+	// At simTimeDelta=1 this means 1 second
+	// of real time = 1 minute of sim time.
+	//If the processor is not fast enough to maintain the specified rate, the frame rate will not be achieved 
 
-		//======= set up Jason BDI agents ================
-		Map<String,AgentModel> agentList = new HashMap<String,AgentModel>();
-		for(UAS uas: UAS_list) {//Create UAS agents
-			agentList.put(uas.getID(), uas);
-		}  
-		jasonAgents = new JasonMAS(agentList);
-		jasonAgents.startAgents();
-		//==========================================
+	//======= set up Jason BDI agents ================
+	Map<String,AgentModel> agentList = new HashMap<String,AgentModel>();
+	for(UAS uas: UAS_list) {//Create UAS agents
+		agentList.put(uas.getID(), uas);
+	}  
+	jasonAgents = new JasonMAS(agentList);
+	jasonAgents.startAgents();
+	//==========================================
 
-	}
+}
 
-	/************* Main draw() ***********************/
-	// Main state update and visualization function
-	// called by Processing in an infinite loop
-	//************************************************/
+/************* Main draw() ***********************/
+// Main state update and visualization function
+// called by Processing in an infinite loop
+//************************************************/
 	public void draw(){
 		if(simPaused){
-
+			
 			background(240); // white background
 
 			for(int i = 0; i < NUMBER_UAS; i++){ //Draw UAS agents
@@ -181,11 +181,11 @@ public void setup() {
 
 			for(int i = 0; i < objects.size(); i++){ //Makes all trees on screen.
 			  drawObject(objects.get(i));
-      }
+			}
 		
-		  for(int i = 0; i < NUMBER_THREATS; i++){ //Makes all trees on screen.
-		    drawThreat(threats.get(i));
-      }
+			for(int i = 0; i < NUMBER_THREATS; i++){ //Makes all trees on screen.
+				drawThreat(threats.get(i));
+			}
       
 			playButton.label="play";
 
@@ -212,72 +212,71 @@ public void setup() {
 		drawUAS(UAS_list.get(i));	
 	}  
     
-  for(int i = 0; i < objects.size(); i++){ //Draw trees.
-    	drawObject(objects.get(i));
-  }
+	for(int i = 0; i < objects.size(); i++){ //Draw trees.
+		drawObject(objects.get(i));
+	}
     
-  for(int i = 0; i < NUMBER_THREATS; i++) {//Draw Threats.
-	  	drawThreat(threats.get(i));
-  }
+	for(int i = 0; i < NUMBER_THREATS; i++) {//Draw Threats.
+		drawThreat(threats.get(i));
+	}
     
-  playButton.label="pause";
+	playButton.label="pause";
 	
 	playButton.drawButton();
 	stopButton.drawButton();
 }
 
 
-	public void mousePressed(){
+public void mousePressed(){
 
-		if(playButton.MouseIsOver()){
-			pauseSimulation();
-		}
-
-		if(stopButton.MouseIsOver()){
-			resetSimulation();
-		}
+	if(playButton.MouseIsOver()){
+		pauseSimulation();
 	}
 
-
-
-	// Visualize
-	public void drawUAS(UAS uas){
-
-		//PVector uasposition, double compassAngle	
-		PVector p1;
-		// Draw UAS
-		stroke(0); 	 
-		PShape s;
-		s=loadShape("robot.svg");
-
-		// translate to center image on uasposition.x, uasposition.y
-		s.translate(-s.width/2,-s.height/2);
-
-		// to adjust compassAngle to the image
-		s.rotate((float) ((float)uas.getCompassAngle()+Math.PI/2));
-
-		//draw image
-		shape(s, uas.getPosition().x, uas.getPosition().y, 26, 26);
-
-		noFill();
-
-		//draw perception area
-		arc(uas.getPosition().x, uas.getPosition().y, PERCEPTION_DISTANCE*2, PERCEPTION_DISTANCE*2,(float)uas.getCompassAngle()-(float)Math.PI/2, (float)uas.getCompassAngle()+(float)Math.PI/2);
-
-		//draw circle on objects percepted
-		ArrayList<VisibleItem> items = new ArrayList<VisibleItem>();
-
-		items = uas.agentState.getCameraInfo(); 
-
-		for(int i=0; i< items.size(); i++) {  
-			double angle = (uas.getCompassAngle()+items.get(i).getAngle());// % 2* Math.PI;
-			double cosv = Math.cos(angle);
-			double sinv = Math.sin(angle);
-			p1 = new PVector(Math.round(cosv*items.get(i).getDistance())+uas.getPosition().x, Math.round(sinv*items.get(i).getDistance())+uas.getPosition().y); 
-			// draw circle over items visualized
-			ellipse(p1.x,p1.y, 26, 26);
-		}
+	if(stopButton.MouseIsOver()){
+		resetSimulation();
 	}
+}
+
+
+// Visualize
+public void drawUAS(UAS uas){
+
+	//PVector uasposition, double compassAngle	
+	PVector p1;
+	// Draw UAS
+	stroke(0); 	 
+	PShape s;
+	s=loadShape("robot.svg");
+
+	// translate to center image on uasposition.x, uasposition.y
+	s.translate(-s.width/2,-s.height/2);
+
+	// to adjust compassAngle to the image
+	s.rotate((float) ((float)uas.getCompassAngle()+Math.PI/2));
+
+	//draw image
+	shape(s, uas.getPosition().x, uas.getPosition().y, 26, 26);
+
+	noFill();
+
+	//draw perception area
+	arc(uas.getPosition().x, uas.getPosition().y, PERCEPTION_DISTANCE*2, PERCEPTION_DISTANCE*2,(float)uas.getCompassAngle()-(float)Math.PI/2, (float)uas.getCompassAngle()+(float)Math.PI/2);
+
+	//draw circle on objects percepted
+	ArrayList<VisibleItem> items = new ArrayList<VisibleItem>();
+
+	items = uas.agentState.getCameraInfo(); 
+
+	for(int i=0; i< items.size(); i++) {  
+		double angle = (uas.getCompassAngle()+items.get(i).getAngle());// % 2* Math.PI;
+		double cosv = Math.cos(angle);
+		double sinv = Math.sin(angle);
+		p1 = new PVector(Math.round(cosv*items.get(i).getDistance())+uas.getPosition().x, Math.round(sinv*items.get(i).getDistance())+uas.getPosition().y); 
+		// draw circle over items visualized
+		ellipse(p1.x,p1.y, 26, 26);
+	}
+}
 
 
 //Visualize
@@ -333,11 +332,11 @@ public void pauseSimulation(){
 }
 
 
-	//************** USER INPUT *****************************/
-	// These functions handle user input events
-	// See "Input" subsection in processing.org/reference
-	//************r*******************************************/
-	public void keyPressed(){  // handle keyboard input
+//************** USER INPUT *****************************/
+// These functions handle user input events
+// See "Input" subsection in processing.org/reference
+//************r*******************************************/
+public void keyPressed(){  // handle keyboard input
 		switch(key) {
 		case 'r': case 'R': resetSimulation(); break; // reset the simulation
 		case ' ':           pauseSimulation(); break; // pause the simulation   
@@ -346,48 +345,48 @@ public void pauseSimulation(){
 	}
 
 
-	//the Button class
-	public class Button{
-		String label; // button label
-		float x;      // top left corner x position
-		float y;      // top left corner y position
-		float w;      // width of button
-		float h;      // height of button
+//the Button class
+public class Button{
+	String label; // button label
+	float x;      // top left corner x position
+	float y;      // top left corner y position
+	float w;      // width of button
+	float h;      // height of button
 
-		// constructor
-		Button(String labelB, float xpos, float ypos, float widthB, float heightB){
-			label = labelB;
-			x = xpos;
-			y = ypos;
-			w = widthB;
-			h = heightB;
-		}
-
-		void drawButton(){
-			shapeMode(CORNER);
-
-			if(label.contentEquals("play")){
-				//s=loadShape("play.svg");
-				shape(play, x, y, w, h);
-			}
-			else if(label.contentEquals("restart")){
-				//s=loadShape("replay.svg");
-				shape(restart, x, y, w, h);
-			}
-			else {
-				//s=loadShape("pause.svg");
-				shape(pause, x, y, w, h);
-			}
-		}
-
-		boolean MouseIsOver(){
-			if (mouseX > x && mouseX < (x + w) && mouseY > y && mouseY < (y + h)){
-				return true;
-			}
-			return false;
-		}
-
+	// constructor
+	Button(String labelB, float xpos, float ypos, float widthB, float heightB){
+		label = labelB;
+		x = xpos;
+		y = ypos;
+		w = widthB;
+		h = heightB;
 	}
+
+	void drawButton(){
+		shapeMode(CORNER);
+
+		if(label.contentEquals("play")){
+			//s=loadShape("play.svg");
+			shape(play, x, y, w, h);
+		}
+		else if(label.contentEquals("restart")){
+			//s=loadShape("replay.svg");
+			shape(restart, x, y, w, h);
+		}
+		else {
+			//s=loadShape("pause.svg");
+			shape(pause, x, y, w, h);
+		}
+	}
+
+	boolean MouseIsOver(){
+		if (mouseX > x && mouseX < (x + w) && mouseY > y && mouseY < (y + h)){
+			return true;
+		}
+		return false;
+	}
+
+}
 
 
 }
